apps/cli/pyproject.toml
[project]
name = "classifyre-cli"
version = "0.2.24"
description = "Classifyre CLI — scan and classify unstructured data sources"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
keywords = ["data", "metadata", "ingestion", "unstructured", "pii", "secrets"]
dependencies = [
    "classifyre-schemas",
    "jsonschema>=4.26.0",
    "pydantic>=2.12.5",
    "email-validator>=2.3.0",
    "requests>=2.32.3",
    "en-core-web-sm",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "crawl4ai>=0.8.0",
]

[tool.uv.sources]
# Local editable install for development — ignored when published to PyPI
classifyre-schemas = { path = "../../packages/schemas", editable = true }
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
de-core-news-lg = { url = "https://github.com/explosion/spacy-models/releases/download/de_core_news_lg-3.8.0/de_core_news_lg-3.8.0-py3-none-any.whl" }

[project.scripts]
classifyre = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "datamodel-code-generator>=0.45.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
]
privacy = [
    "presidio-analyzer>=2.2.361",
    "presidio-anonymizer>=2.2.361",
    "spacy>=3.8.11",
    "scrubadub>=2.0.1",
    "phonenumbers>=9.0.25",
    "pycanon>=1.1.0",
    "commonregex>=1.5.4",
    "pytesseract>=0.3.13",
    "pillow>=12.1.1",
    "de-core-news-lg",
]
security = [
    "detect-secrets>=1.5.0",
    "bandit>=1.9.4",
    "yara-python>=4.5.4",
    "transformers>=4.50.0,<6.0.0",
    "onnxruntime>=1.24.2",
    "optimum[onnxruntime]>=2.1.0",
    "sentencepiece>=0.2.1",
    "torch>=2.10.0",
]
content = [
    "detoxify>=0.5.2",
    "nudenet>=3.4.2",
    "transformers>=4.50.0,<6.0.0",
    "onnxruntime>=1.24.2",
    "torch>=2.10.0",
    "pillow>=12.1.1",
    "sentencepiece>=0.2.1",
]
quality = [
    "fast-langdetect>=1.0.0",
    "datasketch>=1.9.0",
    "sentence-transformers>=5.2.3",
    "textstat>=0.7.13",
    "transformers>=4.50.0,<6.0.0",
    "torch>=2.10.0",
    "sentencepiece>=0.2.1",
]
classification = [
    "transformers>=4.50.0,<6.0.0",
    "fast-langdetect>=1.0.0",
]
custom = [
    "transformers>=4.50.0,<6.0.0",
    "torch>=2.10.0",
    "setfit>=1.1.3",
    "datasets>=4.4.1",
    "scikit-learn>=1.7.2",
    "gliner>=0.2.22",
    "sentence-transformers>=5.2.3",
]
detectors = [
    { include-group = "privacy" },
    { include-group = "security" },
    { include-group = "content" },
    { include-group = "quality" },
    { include-group = "classification" },
    { include-group = "custom" },
]
file-processing = [
    "filetype>=1.2.0",
    "pdfplumber>=0.11.9",
    "python-docx>=1.2.0",
    "openpyxl>=3.1.5",
    "chardet>=5.2.0",
    "pyarrow>=18.0.0",
]
postgresql = [
    "psycopg2-binary>=2.9.9,<3.0.0",
]
mysql = [
    "pymysql>=1.1.1,<2.0.0",
]
mssql = [
    "pymssql>=2.3.4,<3.0.0",
]
oracle = [
    "oracledb>=2.5.1,<3.0.0",
]
hive = [
    "pyhive>=0.7.0,<1.0.0",
    "thrift>=0.16.0,<1.0.0",
    "thrift-sasl>=0.4.3,<1.0.0",
    "pure-sasl>=0.6.2,<1.0.0",
]
databricks = [
    "databricks-sql-connector>=4.0.0,<5.0.0",
]
snowflake = [
    "snowflake-connector-python>=4.0.0,<5.0.0",
]
mongodb = [
    "pymongo>=4.15.2,<5.0.0",
]
tableau = [
    "tableauserverclient>=0.38,<1.0.0",
]
s3-compatible-storage = [
    "boto3>=1.40.40,<2.0.0",
]
azure-blob-storage = [
    "azure-storage-blob>=12.26.0,<13.0.0",
    "azure-identity>=1.25.1,<2.0.0",
]
google-cloud-storage = [
    "google-cloud-storage>=3.4.0,<4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "ARG",  # flake8-unused-arguments
    "PTH",  # flake8-use-pathlib
    "ERA",  # eradicate
    "PL",   # pylint
    "RUF",  # ruff-specific rules
]
ignore = [
    "PLR0913", # too many arguments in function definition
    "B027",    # empty method in abstract base class (used for optional hooks)
    "PLR0912", # too many branches
    "PLR0915", # too many statements
    "E501",    # line too long
    "ERA001",  # eradicate commented-out code
    "UP007",   # Use `X | Y` for type annotations
    "ARG002",  # unused method argument (common in abstract implementations)
    "PLR2004", # magic value in comparison (acceptable in tests and configs)
    "PLR0911", # too many return statements (acceptable for classification methods)
    "PLC0415", # import outside top-level (needed for lazy loading)
    "PTH123",  # open() usage (acceptable in tests)
]

[tool.ruff.lint.isort]
known-first-party = ["src", "schemas"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
show_error_codes = true
plugins = ["pydantic.mypy"]
exclude = ["scripts/"]

[[tool.mypy.overrides]]
module = ["jsonschema.*", "requests.*", "schemas.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["bs4.*", "lxml.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "detect_secrets.*",
    "presidio_analyzer.*",
    "presidio_anonymizer.*",
    "spacy.*",
    "phonenumbers.*",
    "detoxify.*",
    "transformers.*",
    "torch.*",
    "yara.*",
    "crawl4ai.*",
    "nudenet.*",
    "fast_langdetect.*",
    "bandit.*",
    "datasketch.*",
    "sentence_transformers.*",
    "pytesseract.*",
    "optimum.*",
    "onnxruntime.*",
    "timm.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["filetype.*", "pdfplumber.*", "docx.*", "openpyxl.*", "chardet.*", "pyarrow.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "src.models.generated_input",
    "src.models.generated_detectors",
    "src.models.generated_single_asset_scan_results",
    "scripts.*",
]
# Generated files — suppress all mypy errors
ignore_errors = true

[[tool.mypy.overrides]]
module = ["tests.*"]
# Tests don't need full strict typing
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "src.detectors.*",
    "src.models.detectors.*",
    "src.sources.*",
    "src.pipeline.*",
    "src.utils.*",
    "src.outputs.*",
]
# Disable remaining false positives after pydantic plugin
disable_error_code = ["call-arg", "assignment", "no-untyped-def", "no-untyped-call", "no-any-return", "union-attr", "attr-defined", "arg-type", "misc", "list-item", "comparison-overlap", "type-arg", "return-value", "dict-item", "unused-ignore"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests (requires network access)",
    "e2e: marks tests as end-to-end tests (requires real APIs)",
]

[tool.uv]
package = true
